embedding/clip_embedding/requirements.txt
torch
clip@git+https://github.com/openai/CLIP.git
pillow